+++ pyproject.toml
@@ -25,7 +25,7 @@ packages = [
 ]
 
 [tool.poetry.scripts]
-yml2block = 'yml2block.yml2block:main'
+yml2block = 'yml2block.__main__:main'
 
 [tool.poetry.dependencies]
 python = "^3.10"
